+++ packages/backend/src/main.ts
@@ -1,4 +1,4 @@
-import { readFile } from 'fs/promises';
+import { readFile, rm } from 'fs/promises';
 import { existsSync, watch } from 'fs';
 import { SourcebotConfigurationSchema } from "./schemas/v2.js";
 import { getGitHubReposFromConfig } from "./github.js";
@@ -6,7 +6,7 @@ import { getGitLabReposFromConfig } from "./gitlab.js";
 import { getGiteaReposFromConfig } from "./gitea.js";
 import { getGerritReposFromConfig } from "./gerrit.js";
 import { AppContext, LocalRepository, GitRepository, Repository, Settings } from "./types.js";
-import { cloneRepository, fetchRepository } from "./git.js";
+import { cloneRepository, fetchRepository, getGitRepoFromConfig } from "./git.js";
 import { createLogger } from "./logger.js";
 import { createRepository, Database, loadDB, updateRepository, updateSettings } from './db.js';
 import { arraysEqualShallow, isRemotePath, measure } from "./utils.js";
@@ -16,6 +16,8 @@ import stripJsonComments from 'strip-json-comments';
 import { indexGitRepository, indexLocalRepository } from "./zoekt.js";
 import { getLocalRepoFromConfig, initLocalRepoFileWatchers } from "./local.js";
 import { captureEvent } from "./posthog.js";
+import { glob } from 'glob';
+import path from 'path';
 
 const logger = createLogger('main');
 
@@ -66,6 +68,76 @@ const syncLocalRepository = async (repo: LocalRepository, settings: Settings, ct
     return {
         indexDuration_s,
     }
+}
+
+export const deleteStaleRepository = async (repo: Repository, db: Database, ctx: AppContext) => {
+    logger.info(`Deleting stale repository ${repo.id}:`);
+
+    // Delete the checked out git repository (if applicable)
+    if (repo.vcs === "git" && existsSync(repo.path)) {
+        logger.info(`\tDeleting git directory ${repo.path}...`);
+        await rm(repo.path, {
+            recursive: true,
+        });
+    }
+
+    // Delete all .zoekt index files
+    {
+        // .zoekt index files are named with the repository name,
+        // index version, and shard number. Some examples:
+        //
+        //   git repos:
+        //   github.com%2Fsourcebot-dev%2Fsourcebot_v16.00000.zoekt
+        //   gitlab.com%2Fmy-org%2Fmy-project.00000.zoekt
+        //
+        //   local repos:
+        //   UnrealEngine_v16.00000.zoekt
+        //   UnrealEngine_v16.00001.zoekt
+        //   ...
+        //   UnrealEngine_v16.00016.zoekt
+        //
+        // Notice that local repos are named with the repository basename and
+        // git repos are named with the query-encoded repository name. Form a
+        // glob pattern with the correct prefix & suffix to match the correct
+        // index file(s) for the repository.
+        //
+        // @see : https://github.com/sourcegraph/zoekt/blob/c03b77fbf18b76904c0e061f10f46597eedd7b14/build/builder.go#L348
+        const indexFilesGlobPattern = (() => {
+            switch (repo.vcs) {
+                case 'git':
+                    return `${encodeURIComponent(repo.id)}*.zoekt`;
+                case 'local':
+                    return `${path.basename(repo.path)}*.zoekt`;
+            }
+        })();
+
+        const indexFiles = await glob(indexFilesGlobPattern, {
+            cwd: ctx.indexPath,
+            absolute: true
+        });
+
+        await Promise.all(indexFiles.map((file) => {
+            if (!existsSync(file)) {
+                return;
+            }
+
+            logger.info(`\tDeleting index file ${file}...`);
+            return rm(file);
+        }));
+    }
+
+    // Delete db entry
+    logger.info(`\tDeleting db entry...`);
+    await db.update(({ repos }) => {
+        delete repos[repo.id];
+    });
+    
+    logger.info(`Deleted stale repository ${repo.id}`);
+
+    captureEvent('repo_deleted', {
+        vcs: repo.vcs,
+        codeHost: repo.codeHost,
+    })
 }
 
 /**
@@ -138,6 +210,10 @@ const syncConfig = async (configPath: string, db: Database, signal: AbortSignal,
     // Update the settings
     const updatedSettings: Settings = {
         maxFileSize: config.settings?.maxFileSize ?? DEFAULT_SETTINGS.maxFileSize,
+        maxTrigramCount: config.settings?.maxTrigramCount ?? DEFAULT_SETTINGS.maxTrigramCount,
+        autoDeleteStaleRepos: config.settings?.autoDeleteStaleRepos ?? DEFAULT_SETTINGS.autoDeleteStaleRepos,
+        reindexInterval: config.settings?.reindexInterval ?? DEFAULT_SETTINGS.reindexInterval,
+        resyncInterval: config.settings?.resyncInterval ?? DEFAULT_SETTINGS.resyncInterval,
     }
     const _isAllRepoReindexingRequired = isAllRepoReindexingRequired(db.data.settings, updatedSettings);
     await updateSettings(updatedSettings, db);
@@ -169,6 +245,11 @@ const syncConfig = async (configPath: string, db: Database, signal: AbortSignal,
             case 'local': {
                 const repo = getLocalRepoFromConfig(repoConfig, ctx);
                 configRepos.push(repo);
+                break;
+            }
+            case 'git': {
+                const gitRepo = await  getGitRepoFromConfig(repoConfig, ctx);
+                gitRepo && configRepos.push(gitRepo);
                 break;
             }
         }
@@ -278,11 +359,10 @@ export const main = async (context: AppContext) => {
         });
     }
 
-    // Re-sync every 24 hours
+    // Re-sync at a fixed interval
     setInterval(() => {
-        logger.info(`Re-syncing configuration file ${context.configPath}`);
         _syncConfig();
-    }, RESYNC_CONFIG_INTERVAL_MS);
+    }, db.data.settings.resyncInterval);
 
     // Sync immediately on startup
     await _syncConfig();
@@ -293,10 +373,16 @@ export const main = async (context: AppContext) => {
         for (const [_, repo] of Object.entries(repos)) {
             const lastIndexed = repo.lastIndexedDate ? new Date(repo.lastIndexedDate) : new Date(0);
 
-            if (
-                repo.isStale ||
-                lastIndexed.getTime() > Date.now() - REINDEX_INTERVAL_MS
-            ) {
+            if (repo.isStale) {
+                if (db.data.settings.autoDeleteStaleRepos) {
+                    await deleteStaleRepository(repo, db, context);
+                } else {
+                    // skip deletion...
+                }
+                continue;
+            }
+
+            if (lastIndexed.getTime() > (Date.now() - db.data.settings.reindexInterval)) {
                 continue;
             }