+++ packages/web/src/app/api/(client)/client.ts
@@ -1,8 +1,9 @@
 'use client';
 
 import { NEXT_PUBLIC_DOMAIN_SUB_PATH } from "@/lib/environment.client";
-import { fileSourceResponseSchema, listRepositoriesResponseSchema, searchResponseSchema } from "@/lib/schemas";
-import { FileSourceRequest, FileSourceResponse, ListRepositoriesResponse, SearchRequest, SearchResponse } from "@/lib/types";
+import { fileSourceResponseSchema, getVersionResponseSchema, listRepositoriesResponseSchema, searchResponseSchema } from "@/lib/schemas";
+import { FileSourceRequest, FileSourceResponse, GetVersionResponse, ListRepositoriesResponse, SearchRequest, SearchResponse } from "@/lib/types";
+import assert from "assert";
 
 export const search = async (body: SearchRequest): Promise<SearchResponse> => {
     const path = resolveServerPath("/api/search");
@@ -42,11 +43,23 @@ export const getRepos = async (): Promise<ListRepositoriesResponse> => {
     return listRepositoriesResponseSchema.parse(result);
 }
 
+export const getVersion = async (): Promise<GetVersionResponse> => {
+    const path = resolveServerPath("/api/version");
+    const result = await fetch(path, {
+        method: "GET",
+        headers: {
+            "Content-Type": "application/json",
+        },
+    }).then(response => response.json());
+    return getVersionResponseSchema.parse(result);
+}
+
 /**
  * Given a subpath to a api route on the server (e.g., /api/search),
  * returns the full path to that route on the server, taking into account
  * the base path (if any).
  */
 export const resolveServerPath = (path: string) => {
+    assert(path.startsWith("/"));
     return `${NEXT_PUBLIC_DOMAIN_SUB_PATH}${path}`;
 }